added distribution to RVS and removed duplication
