+++ cabby/model/text/dual_encoder/dataset_run.py
@@ -20,7 +20,6 @@ import os
 import pandas as pd
 from sklearn.utils import shuffle
 import torch
-from transformers import DistilBertTokenizerFast
 
 from cabby.geo import util as gutil
 from cabby.model.text import util 
@@ -28,102 +27,12 @@ from cabby.model.text import util
 from cabby.geo import regions
 from cabby.model.text.dual_encoder import dataset_item
 from cabby.model import datasets
+
 from cabby.model import util as mutil
-dprob = mutil.DistanceProbability(37)
 
-
-tokenizer = DistilBertTokenizerFast.from_pretrained('distilbert-base-uncased')
-
-CELLID_DIM = 64
-
-
-class TextGeoSplit(torch.utils.data.Dataset):
-  """A split of of the RUN dataset.
-  
-  `points`: The ground true end-points of the samples.
-  `labels`: The ground true label of the cellid.
-  `cellids`: The ground truth S2Cell id.
-  `neighbor_cells`: One neighbor cell id of the ground truth S2Cell id.
-  `far_cells`: One far away cell id (in the region defined) of the ground truth 
-  S2Cell id.
-  """
-  def __init__(self, data: pd.DataFrame, s2level: int, 
-    unique_cells_df: pd.DataFrame, cellid_to_label: Dict[int, int]):
-    # Tokenize instructions.
-    self.encodings = tokenizer(
-      data.instructions.tolist(), truncation=True,
-      padding=True, add_special_tokens=True)
-
-    start_points = data.start_point.apply(
-      lambda x: gutil.point_from_list_coord(x))
-    
-    dist_lists = start_points.apply(
-      lambda start: unique_cells_df.point.apply(
-        lambda end: gutil.get_distance_between_points(start, end))
-        ).values.tolist()
-    
-
-    self.prob = []
-    for dist_list in dist_lists:
-      prob = []
-      for dist in dist_list:
-        prob.append(dprob(dist))
-      self.prob.append(prob)
-
-    points = data.end_point.apply(
-      lambda x: gutil.point_from_list_coord(x))
-
-    data = data.assign(point=points)
-
-    data['cellid'] = data.point.apply(
-      lambda x: gutil.cellid_from_point(x, s2level))
-
-    data['neighbor_cells'] = data.cellid.apply(
-      lambda x: gutil.neighbor_cellid(x))
-
-    data['far_cells'] = data.cellid.apply(
-      lambda cellid: unique_cells_df[unique_cells_df['cellid']==cellid].far.iloc[0])
-
-    cellids_array = np.array(data.cellid.tolist())
-    neighbor_cells_array = np.array(data.neighbor_cells.tolist())
-    far_cells_array = np.array(data.far_cells.tolist())
-
-
-    self.points = data.point.apply(
-      lambda x: gutil.tuple_from_point(x)).tolist()
-    self.labels = data.cellid.apply(lambda x: cellid_to_label[x]).tolist()
-    self.cellids = util.binary_representation(cellids_array, dim = CELLID_DIM)
-    self.neighbor_cells =  util.binary_representation(
-      neighbor_cells_array, dim = CELLID_DIM)
-    self.far_cells =  util.binary_representation(
-      far_cells_array, dim = CELLID_DIM)
-
-  def __getitem__(self, idx: int):
-    '''Supports indexing such that TextGeoDataset[i] can be used to get 
-    i-th sample. 
-    Arguments:
-      idx: The index for which a sample from the dataset will be returned.
-    Returns:
-      A single sample including text, the correct cellid, a neighbor cellid, 
-      a far cellid, a point of the cellid and the label of the cellid.
-    '''
-    text = {key: torch.tensor(val[idx])
-        for key, val in self.encodings.items()}
-    cellid = torch.tensor(self.cellids[idx])
-    neighbor_cells = torch.tensor(self.neighbor_cells[idx])
-    far_cells = torch.tensor(self.far_cells[idx])
-    point = torch.tensor(self.points[idx])
-    label = torch.tensor(self.labels[idx])
-    prob = torch.tensor(self.prob[idx])
-    
-    sample = {'text': text, 'cellid': cellid, 'neighbor_cells': neighbor_cells, 
-      'far_cells': far_cells, 'point': point, 'label': label, 'prob': prob}
-
-    return sample
-
-  def __len__(self):
-    return len(self.cellids)
-
+# DISTRIBUTION_SCALE_DISTANCEA is a factor (in meters) that gives the overall scale in meters for the distribution.
+DISTRIBUTION_SCALE_DISTANCE = 37
+dprob = mutil.DistanceProbability(DISTRIBUTION_SCALE_DISTANCE)
 
 def create_dataset(
                   data_dir: str, 
@@ -150,24 +59,27 @@ def create_dataset(
       lambda x: gutil.far_cellid(x, unique_cells_df))
 
   vec_cells = util.binary_representation(unique_cells_df.cellid.to_numpy(), 
-  dim = CELLID_DIM)
+  dim = dataset_item.CELLID_DIM)
   tens_cells = torch.tensor(vec_cells)
 
-  # Create RVS dataset.
+  # Create RUN dataset.
   train_dataset = None
   val_dataset = None
   logging.info("Starting to create the splits")
   if infer_only == False:
-    train_dataset = TextGeoSplit(
-      run_dataset.train, s2level, unique_cells_df, run_dataset.cellid_to_label)
+    train_dataset = dataset_item.TextGeoSplit(
+      run_dataset.train, s2level, unique_cells_df, 
+      run_dataset.cellid_to_label, dprob)
     logging.info(
       f"Finished to create the train-set with {len(train_dataset)} samples")
-    val_dataset = TextGeoSplit(
-      run_dataset.valid, s2level, unique_cells_df, run_dataset.cellid_to_label)
+    val_dataset = dataset_item.TextGeoSplit(
+      run_dataset.valid, s2level, unique_cells_df, 
+      run_dataset.cellid_to_label, dprob)
     logging.info(
       f"Finished to create the valid-set with {len(val_dataset)} samples")
-  test_dataset = TextGeoSplit(
-    run_dataset.test, s2level, unique_cells_df, run_dataset.cellid_to_label)
+  test_dataset = dataset_item.TextGeoSplit(
+    run_dataset.test, s2level, unique_cells_df, 
+    run_dataset.cellid_to_label, dprob)
   logging.info(
     f"Finished to create the test-set with {len(test_dataset)} samples")
