+++ cabby/model/text/dual_encoder/dataset_rvs.py
@@ -20,10 +20,8 @@ import os
 import pandas as pd
 
 import torch
-from transformers import DistilBertTokenizerFast
 
 from cabby.geo import util as gutil
-from cabby.model import util as mutil
 from cabby.model.text import util 
 
 
@@ -31,83 +29,9 @@ from cabby.geo import regions
 from cabby.model.text.dual_encoder import dataset_item
 from cabby.model import datasets
 
-
-tokenizer = DistilBertTokenizerFast.from_pretrained('distilbert-base-uncased')
-
-CELLID_DIM = 64
-
-class TextGeoSplit(torch.utils.data.Dataset):
-  """A split of of the RVS dataset.
-  
-  `points`: The ground true end-points of the samples.
-  `labels`: The ground true label of the cellid.
-  `cellids`: The ground truth S2Cell id.
-  `neighbor_cells`: One neighbor cell id of the ground truth S2Cell id.
-  `far_cells`: One far away cell id (in the region defined) of the ground truth 
-  S2Cell id.
-  """
-  def __init__(self, data: pd.DataFrame, s2level: int, 
-    unique_cells_df: pd.DataFrame, cellid_to_label: Dict[int, int]):
-    # Tokenize instructions.
-  
-    self.encodings = tokenizer(
-      data.instructions.tolist(), truncation=True,
-      padding=True, add_special_tokens=True)
-
-    end_points = data.end_point.apply(
-      lambda x: gutil.point_from_list_coord(x))
-
-    data = data.assign(end_point=end_points)
-
-    data['cellid'] = data.end_point.apply(
-      lambda x: gutil.cellid_from_point(x, s2level))
-
-    data['neighbor_cells'] = data.cellid.apply(
-      lambda x: gutil.neighbor_cellid(x))
-
-    data['far_cells'] = data.cellid.apply(
-      lambda cellid: unique_cells_df[unique_cells_df['cellid']==cellid].far.iloc[0])
-
-    cellids_array = np.array(data.cellid.tolist())
-    neighbor_cells_array = np.array(data.neighbor_cells.tolist())
-    far_cells_array = np.array(data.far_cells.tolist())
-
-
-    self.end_points = data.end_point.apply(
-      lambda x: gutil.tuple_from_point(x)).tolist()
-    self.labels = data.cellid.apply(lambda x: cellid_to_label[x]).tolist()
-    self.cellids = util.binary_representation(cellids_array, dim = CELLID_DIM)
-    self.neighbor_cells =  util.binary_representation(
-      neighbor_cells_array, dim = CELLID_DIM)
-    self.far_cells =  util.binary_representation(
-      far_cells_array, dim = CELLID_DIM)
-
-    
-  def __getitem__(self, idx: int):
-    '''Supports indexing such that TextGeoDataset[i] can be used to get 
-    i-th sample. 
-    Arguments:
-      idx: The index for which a sample from the dataset will be returned.
-    Returns:
-      A single sample including text, the correct cellid, a neighbor cellid, 
-      a far cellid, a point of the cellid and the label of the cellid.
-    '''
-    text = {key: torch.tensor(val[idx])
-        for key, val in self.encodings.items()}
-    cellid = torch.tensor(self.cellids[idx])
-    neighbor_cells = torch.tensor(self.neighbor_cells[idx])
-    far_cells = torch.tensor(self.far_cells[idx])
-    end_point = torch.tensor(self.end_points[idx])
-    label = torch.tensor(self.labels[idx])
-    
-    sample = {'text': text, 'cellid': cellid, 'neighbor_cells': neighbor_cells, 
-      'far_cells': far_cells, 'point': end_point, 'label': label, }
-
-    return sample
-
-  def __len__(self):
-    return len(self.cellids)
-
+# DISTRIBUTION_SCALE_DISTANCEA is a factor (in meters) that gives the overall scale in meters for the distribution.
+DISTRIBUTION_SCALE_DISTANCE = 1000
+dprob = mutil.DistanceProbability(DISTRIBUTION_SCALE_DISTANCE)
 
 def create_dataset(
                   data_dir: str, 
@@ -133,7 +57,7 @@ def create_dataset(
       lambda x: gutil.far_cellid(x, unique_cells_df))
 
   vec_cells = util.binary_representation(unique_cells_df.cellid.to_numpy(), 
-  dim = CELLID_DIM)
+  dim = dataset_item.CELLID_DIM)
   tens_cells = torch.tensor(vec_cells)
 
   # Create RVS dataset.
@@ -141,20 +65,23 @@ def create_dataset(
   val_dataset = None
   logging.info("Starting to create the splits")
   if infer_only == False:
-    train_dataset = TextGeoSplit(
-      rvs_dataset.train, s2level, unique_cells_df, rvs_dataset.cellid_to_label)
+    train_dataset = dataset_item.TextGeoSplit(
+      run_dataset.train, s2level, unique_cells_df, 
+      run_dataset.cellid_to_label, dprob)
     logging.info(
       f"Finished to create the train-set with {len(train_dataset)} samples")
-    val_dataset = TextGeoSplit(
-      rvs_dataset.valid, s2level, unique_cells_df, rvs_dataset.cellid_to_label)
+    val_dataset = dataset_item.TextGeoSplit(
+      run_dataset.valid, s2level, unique_cells_df, 
+      run_dataset.cellid_to_label, dprob)
     logging.info(
       f"Finished to create the valid-set with {len(val_dataset)} samples")
-  test_dataset = TextGeoSplit(
-    rvs_dataset.test, s2level, unique_cells_df, rvs_dataset.cellid_to_label)
+  test_dataset = dataset_item.TextGeoSplit(
+    run_dataset.test, s2level, unique_cells_df, 
+    run_dataset.cellid_to_label, dprob)
   logging.info(
     f"Finished to create the test-set with {len(test_dataset)} samples")
 
   return dataset_item.TextGeoDataset.from_TextGeoSplit(
     train_dataset, val_dataset, test_dataset, 
-    np.array(rvs_dataset.unique_cellid), 
-    tens_cells, rvs_dataset.label_to_cellid)
+    np.array(run_dataset.unique_cellid), 
+    tens_cells, run_dataset.label_to_cellid)
