+++ cabby/model/text/dual_encoder/train.py
@@ -44,6 +44,8 @@ class Trainer:
           num_epochs: int,
           cells_tensor: torch.tensor,
           label_to_cellid: Dict[int, int],
+          is_distance_distribution: bool
+
           ):
 
     self.model = model
@@ -63,7 +65,8 @@ class Trainer:
       os.mkdir(self.file_path)
     self.model_path = os.path.join(self.file_path, 'model.pt')
     self.metrics_path = os.path.join(self.file_path, 'metrics.tsv')
-        
+    self.is_distance_distribution = is_distance_distribution
+    
 
   def evaluate(self, validation_set: bool = True):
     '''Validate the model.'''
@@ -100,8 +103,9 @@ class Trainer:
           batch_dim, cell_dim, output_dim)
         text_embedding_exp = text_embedding.unsqueeze(1)
         output = self.cos(cellid_embedding_exp, text_embedding_exp)
-        prob = batch['prob'].float().to(self.device) 
-        output = output*prob
+        if self.is_distance_distribution:
+          prob = batch['prob'].float().to(self.device) 
+          output = output*prob
         output = output.detach().cpu().numpy()
         predictions = np.argmax(output, axis=1)
         predictions_list.append(predictions)
